+++ torchmetrics/functional/text/infolm.py
@@ -1,19 +1,29 @@
 import os
 from enum import unique
-from typing import Optional, Sequence, Tuple, Union
+from typing import Dict, Optional, Sequence, Tuple, Union
 
 import torch
 from torch import Tensor
+from torch.nn import functional as F
+from torch.utils.data import DataLoader
 from typing_extensions import Literal
 
-from torchmetrics.functional.text.helper_embedding_metric import TextDataset
+from torchmetrics.functional.text.helper_embedding_metric import (
+    TokenizedDataset,
+    _get_progress_bar,
+    _input_data_collator,
+    _load_tokenizer_and_model,
+)
 from torchmetrics.utilities.enums import EnumStr
 from torchmetrics.utilities.imports import _TRANSFORMERS_AVAILABLE
 
 if _TRANSFORMERS_AVAILABLE:
-    from transformers import AutoModelForMaskedLM, AutoTokenizer, PreTrainedModel, PreTrainedTokenizerBase
+    from transformers import PreTrainedModel, PreTrainedTokenizerBase
 else:
     __doctest_skip__ = ["infolm"]
+
+
+_PAD_TOKEN_PROB = 10000
 
 
 _ALLOWED_INFORMATION_MEASURE = (
@@ -278,38 +288,129 @@ class _InformationMeasure:
         return 2 * torch.acos(torch.clamp(torch.sqrt(preds_distribution * target_distribution).sum(-1), 0, 1))
 
 
-def _load_tokenizer_and_model(
-    model_name_or_path: Union[str, os.PathLike], device: Union[str, torch.device]
-) -> Tuple[PreTrainedTokenizerBase, PreTrainedModel]:
-    """Load HuggingFace `transformers`' tokenizer and model. This function also handle a device placement.
+def _get_dataloader(
+    input_ids: Tensor, attention_mask: Tensor, idf: bool, batch_size: int, num_workers: int
+) -> DataLoader:
+    dataset = TokenizedDataset(input_ids, attention_mask, idf)
+    dataloader = DataLoader(dataset, batch_size=batch_size, num_workers=num_workers)
+    return dataloader
 
-    Args:
-        model_name_or_path:
-            A name or a model path used to load `transformers` pretrained model.
-        device:
-            A device to be used for calculation.
 
-    Return:
-        Initialized `transformers`' tokenizer and model.
+def _get_mlm_labels(input_ids: Tensor, mask_idx: int) -> Tensor:
+    mask = torch.zeros_like(input_ids).bool()
+    mask[:, mask_idx] = 1
+    mlm_labels = torch.where(mask, input_ids, -100)
+    return mlm_labels
+
+
+def _get_pad_token_mask(
+    input_ids: Tensor, pad_token_id: int, cls_token_id: int, sep_token_id: int, mask_idx: int
+) -> Tensor:
     """
-    tokenizer = AutoTokenizer.from_pretrained(model_name_or_path)
-    model = AutoModelForMaskedLM.from_pretrained(model_name_or_path, output_hidden_states=True)
-    model.eval()
-    model.to(device)
-    return tokenizer, model
+    Args:
+        input_ids:
+        pad_token_id:
+        cls_token_id:
+        sep_token_id:
+        mask_idx:
+    """
+    mlm_labels = _get_mlm_labels(input_ids, mask_idx)
+    _seq_len, mlm_labels = mlm_labels.shape[-1], mlm_labels[:, mask_idx]
+    pad_token_mask = mlm_labels.eq(pad_token_id) | mlm_labels.eq(cls_token_id) | mlm_labels.eq(sep_token_id)
+    pad_token_mask = pad_token_mask.unsqueeze(1).repeat(1, _seq_len)
+    return pad_token_mask
 
 
-def _infolm_update(preds: Union[str, Sequence[str]], target: Sequence[Union[str, Sequence[str]]]):
-    pass
+def _get_batch_distribution(
+    model: PreTrainedModel, batch: Dict[str, Tensor], temperature: float, max_length: int, idf: bool
+):
+    """
+    Args:
+        model:
+        batch:
+        temperature:
+            A temperature for calibrating language modelling. For more information, please reference `InfoLM`_ paper.
+        max_length:
+        idf:
+            An indication of whether normalization using inverse document frequencies should be used.
+    """
+    seq_len = batch["input_ids"].shape[1]
+    len_to_pad = max_length - seq_len
+    pad_token_id, cls_token_id, sep_token_id = (
+        model.config.pad_token_id,
+        model.config.cls_token_id,
+        model.config.sep_token_id,
+    )
+
+    for mask_idx in range(seq_len):
+        input_ids = batch["input_ids"].clone()
+        pad_token_mask = _get_pad_token_mask(input_ids, pad_token_id, cls_token_id, sep_token_id, mask_idx)
+        input_ids[:, mask_idx] = model.config.mask_token_id
+
+        logits_distribution = model(input_ids, batch["attention_mask"])
+        prob_distribution = F.softmax(logits_distribution / temperature, dim=-1)
+        prob_distribution[pad_token_mask] = _PAD_TOKEN_PROB
+        prob_distribution = torch.cat(
+            [prob_distribution, _PAD_TOKEN_PROB * torch.ones(len_to_pad, prob_distribution.shape[-1])], dim=0
+        )
+
+        del input_ids
 
 
-def _infolm_compute():
-    pass
+@torch.no_grad()
+def _get_data_distribution(
+    model: PreTrainedModel, dataloader: DataLoader, temperature: float, max_length: int, idf: bool, verbose: bool
+):
+    """
+    Args:
+        model:
+        dataloader:
+        temperature:
+        max_length
+        idf:
+        verbose:
+    """
+    device = model.device
+
+    for batch in _get_progress_bar(dataloader, verbose):
+        batch = _input_data_collator(batch, device)
+        _get_batch_distribution(model, batch, temperature, max_length, idf)
+
+
+def _infolm_update(
+    preds: Union[str, Sequence[str]],
+    target: Union[str, Sequence[str]],
+    tokenizer: PreTrainedTokenizerBase,
+    max_length: int,
+) -> Tuple[Tensor, Tensor, Tensor, Tensor]:
+    preds_input = tokenizer(preds, padding="max_length", max_length=max_length, truncation=True, return_tensors="pt")
+    target_input = tokenizer(target, padding="max_length", max_length=max_length, truncation=True, return_tensors="pt")
+
+    return preds_input.input_ids, preds_input.attention_mask, target_input.input_ids, target_input.attention_mask
+
+
+def _infolm_compute(
+    model: PreTrainedModel,
+    preds_dataloader: DataLoader,
+    target_dataloader: DataLoader,
+    temperature: float,
+    information_measure: _InformationMeasure,
+    alpha: Optional[float],
+    beta: Optional[float],
+    max_length: int,
+    verbose: bool = True,
+):
+    """
+    Args:
+    """
+    idf = preds_dataloader.dataset.idf
+    preds_distribution = _get_data_distribution(model, preds_dataloader, temperature, idf, verbose)
+    target_distribution = _get_data_distribution(model, target_dataloader, temperature, idf, verbose)
 
 
 def infolm(
     preds: Union[str, Sequence[str]],
-    target: Sequence[Union[str, Sequence[str]]],
+    target: Union[str, Sequence[str]],
     model_name_or_path: Union[str, os.PathLike] = "bert-base-uncased",
     temperature: float = 0.25,
     information_measure: _ALLOWED_INFORMATION_MEASURE_LITERAL = "kl_divergence",
@@ -317,9 +418,13 @@ def infolm(
     alpha: Optional[float] = None,
     beta: Optional[float] = None,
     device: Optional[Union[str, torch.device]] = None,
-):
+    max_length: Optional[int] = None,
+    batch_size: int = 64,
+    num_threads: int = 4,
+    verbose: bool = True,
+) -> Tensor:
     """
-    Calculate `InfoLM`_ - i.e. calculate a distance/divergence between predicted and reference sentence discrete
+    Calculate `InfoLM`_ [1] - i.e. calculate a distance/divergence between predicted and reference sentence discrete
     distribution using one of the following information measures:
         - `KL divergence`_
         - `alpha divergence`_
@@ -331,11 +436,15 @@ def infolm(
         - L-infinity distance
         - `Fisher-Rao distance`_
 
+    `InfoLM`_ is a family of untrained embedding-based metrics which addresses some famous flaws of standard
+    string-based metrics thanks to the usage of pre-trained masked language models. This family of metrics is mainly
+    designed for summarization and data-to-text tasks.
+
     Args:
         preds:
             An iterable of hypothesis corpus.
         target:
-            An iterable of iterables of reference corpus.
+            An iterable of reference corpus.
         model_name_or_path:
             A name or a model path used to load `transformers` pretrained model.
         temperature:
@@ -352,8 +461,45 @@ def infolm(
             Beta parameter of the divergence used for beta and AB divergence measures.
         device:
             A device to be used for calculation.
+        max_length:
+            A maximum length of input sequences. Sequences longer than `max_length` are to be trimmed.
+        batch_size:
+            A batch size used for model processing.
+        num_threads:
+            A number of threads to use for a dataloader.
+        verbose:
+            An indication of whether a progress bar to be displayed during the embeddings calculation.
+
+    Return:
+
+    Example:
+        >>> from torchmetrics.functional.text.infolm import infolm
+        >>> preds = []
+        >>> target = []
+        >>> infolm(preds, target)
+
+    References:
+        [1] InfoLM: A New Metric to Evaluate Summarization & Data2Text Generation by Pierre Colombo, Chloé Clavel and
+        Pablo Piantanida `InfoLM`_
     """
     tokenizer, model = _load_tokenizer_and_model(model_name_or_path, device)
     information_measure_cls = _InformationMeasure(information_measure, alpha, beta)
+    max_length = max_length or model.config.max_length
 
-    _infolm_update(preds, target)
+    preds_input_ids, preds_attention_mask, target_input_ids, target_attention_mask = _infolm_update(
+        preds, target, tokenizer, max_length
+    )
+    preds_dataloader = _get_dataloader(preds_input_ids, preds_attention_mask, idf, batch_size)
+    target_dataloader = _get_dataloader(target_input_ids, target_attention_mask, idf, batch_size)
+
+    info_lm_score = _infolm_compute(
+        model,
+        preds_dataloader,
+        target_dataloader,
+        temperature,
+        information_measure_cls,
+        alpha,
+        beta,
+    )
+
+    return info_lm_score